fix(keda): expose real admin metrics routes
