+++ backend/api/src/routes/kedaRoutes.js
@@ -1,0 +1,135 @@
+import express from 'express';
+import kedaService from '../services/kedaService.js';
+import logger from '../middleware/logger.js';
+
+const router = express.Router();
+
+router.get('/keda/metrics/requests', async (_req, res) => {
+    try {
+        const result = await kedaService.getAPIRequests();
+        if (!result.success) return res.status(502).json(result);
+        res.json({ success: true, data: result });
+    } catch (error) {
+        logger.error('Requests error:', error);
+        res.status(500).json({ success: false, error: error.message });
+    }
+});
+
+router.get('/keda/metrics/latency', async (_req, res) => {
+    try {
+        const result = await kedaService.getAPILatency();
+        if (!result.success) return res.status(502).json(result);
+        res.json({ success: true, data: result });
+    } catch (error) {
+        logger.error('Latency error:', error);
+        res.status(500).json({ success: false, error: error.message });
+    }
+});
+
+router.get('/keda/metrics/cpu', async (req, res) => {
+    try {
+        const { namespace, deployment } = req.query;
+        if (!namespace || !deployment) {
+            return res.status(400).json({
+                success: false,
+                error: 'namespace and deployment required'
+            });
+        }
+        const result = await kedaService.getCPUUsage(namespace, deployment);
+        if (!result.success) return res.status(502).json(result);
+        res.json({ success: true, data: result });
+    } catch (error) {
+        logger.error('CPU error:', error);
+        res.status(500).json({ success: false, error: error.message });
+    }
+});
+
+router.get('/keda/metrics/memory', async (req, res) => {
+    try {
+        const { namespace, deployment } = req.query;
+        if (!namespace || !deployment) {
+            return res.status(400).json({
+                success: false,
+                error: 'namespace and deployment required'
+            });
+        }
+        const result = await kedaService.getMemoryUsage(namespace, deployment);
+        if (!result.success) return res.status(502).json(result);
+        res.json({ success: true, data: result });
+    } catch (error) {
+        logger.error('Memory error:', error);
+        res.status(500).json({ success: false, error: error.message });
+    }
+});
+
+router.get('/keda/metrics/kafka-lag', async (req, res) => {
+    try {
+        const { topic, consumerGroup } = req.query;
+        if (!topic || !consumerGroup) {
+            return res.status(400).json({
+                success: false,
+                error: 'topic and consumerGroup required'
+            });
+        }
+        const result = await kedaService.getKafkaLag(topic, consumerGroup);
+        if (!result.success) {
+            return res.status(502).json({ success: false, error: result.error || 'Kafka lag metric unavailable' });
+        }
+        res.json({ success: true, data: result });
+    } catch (error) {
+        logger.error('Kafka lag error:', error);
+        res.status(500).json({ success: false, error: error.message });
+    }
+});
+
+router.get('/keda/metrics/autoscale', async (req, res) => {
+    try {
+        const { namespace, deployment } = req.query;
+        if (!namespace || !deployment) {
+            return res.status(400).json({
+                success: false,
+                error: 'namespace and deployment required'
+            });
+        }
+        const result = await kedaService.getAutoscalingMetrics(namespace, deployment);
+        if (!result.success) {
+            return res.status(502).json(result);
+        }
+        res.json({ success: true, data: result });
+    } catch (error) {
+        logger.error('Autoscale metrics error:', error);
+        res.status(500).json({ success: false, error: error.message });
+    }
+});
+
+router.get('/keda/scale/recommend', async (req, res) => {
+    try {
+        const { namespace, deployment } = req.query;
+        if (!namespace || !deployment) {
+            return res.status(400).json({
+                success: false,
+                error: 'namespace and deployment required'
+            });
+        }
+        const result = await kedaService.getScaleRecommendation(namespace, deployment);
+        if (!result.success) {
+            return res.status(502).json(result);
+        }
+        res.json({ success: true, data: result });
+    } catch (error) {
+        logger.error('Scale recommendation error:', error);
+        res.status(500).json({ success: false, error: error.message });
+    }
+});
+
+router.get('/keda/stats', async (_req, res) => {
+    try {
+        const stats = await kedaService.getStats();
+        res.json({ success: true, data: stats });
+    } catch (error) {
+        logger.error('Stats error:', error);
+        res.status(500).json({ success: false, error: error.message });
+    }
+});
+
+export default router;
